configure queues and other details

+++ pen-payment-validator/src/main/java/com/mjp/pen_payment_validator/repositories/PaymentValidatorRepository.java
@@ -9,5 +9,5 @@ import java.util.UUID;
 @Repository
 public interface PaymentValidatorRepository extends JpaRepository<PaymentDetails, UUID> {
 
-    PaymentDetails findByOrderNumber(Long orderNumber);
+    PaymentDetails findByOrderNumber(Integer orderNumber);
 }
